+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## HEAD
 
+* Make logging plugin accept PORO instance which don't have an #id (janko-m)
+
 * Add rack_file plugin for attaching Rack file hashes to models (janko-m)